fix: fix calculation get_S
